+++ src/plugins/views/HomeTab/components/QuickAccess.jsx
@@ -1,18 +1,18 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import PropTypes from "prop-types";
 import { ContextMenu } from "@mov-ai/mov-fe-lib-react";
 import Paper from "@material-ui/core/Paper";
 import Divider from "@material-ui/core/Divider";
 import AddIcon from "@material-ui/icons/Add";
-import BuildIcon from "@material-ui/icons/Build";
 import ChromeReaderModeIcon from "@material-ui/icons/ChromeReaderMode";
 import { getIconByScope } from "../../../../utils/Utils";
-import {
-  APP_DEFAULT_CONFIG,
-  APP_CUSTOM_CONFIG
-} from "../../../../utils/Constants";
-import Configuration from "../../../../models/Configuration/Configuration";
+// This is already working - just leaving this here for posteriority
+// import {
+//   APP_DEFAULT_CONFIG,
+//   APP_CUSTOM_CONFIG
+// } from "../../../../utils/Constants";
+// import Configuration from "../../../../models/Configuration/Configuration";
 
 import { quickAccessStyles } from "../styles";
 
@@ -30,33 +30,34 @@ const QuickAccess = props => {
    *                                                                                      */
   //========================================================================================
 
-  const handleOpenAppConfig = useCallback(() => {
-    const name = APP_CUSTOM_CONFIG;
-    const scope = Configuration.SCOPE;
+  // This is already working - just leaving this here for posteriority
+  // const handleOpenAppConfig = useCallback(() => {
+  //   const name = APP_CUSTOM_CONFIG;
+  //   const scope = Configuration.SCOPE;
 
-    call("docManager", "checkDocumentExists", {
-      name,
-      scope
-    }).then(fileExists => {
-      if (!fileExists) {
-        call(
-          "docManager",
-          "copy",
-          {
-            name: APP_DEFAULT_CONFIG,
-            scope
-          },
-          name
-        );
-      }
+  //   call("docManager", "checkDocumentExists", {
+  //     name,
+  //     scope
+  //   }).then(fileExists => {
+  //     if (!fileExists) {
+  //       call(
+  //         "docManager",
+  //         "copy",
+  //         {
+  //           name: APP_DEFAULT_CONFIG,
+  //           scope
+  //         },
+  //         name
+  //       );
+  //     }
 
-      call("tabs", "openEditor", {
-        id: `global/${scope}/${name}`,
-        name,
-        scope
-      });
-    });
-  }, [call]);
+  //     call("tabs", "openEditor", {
+  //       id: `global/${scope}/${name}`,
+  //       name,
+  //       scope
+  //     });
+  //   });
+  // }, [call]);
 
   //========================================================================================
   /*                                                                                      *
@@ -103,7 +104,7 @@ const QuickAccess = props => {
           }))}
         ></ContextMenu>
         <a
-          href="https://movai.atlassian.net/wiki/spaces/MW/overview"
+          href="https://movai-flow.readme.io/docs"
           target="_blank"
           rel="noreferrer"
           className={classes.link}
@@ -112,10 +113,20 @@ const QuickAccess = props => {
 
           {t("Documentation")}
         </a>
-        <div className={classes.link} onClick={handleOpenAppConfig}>
+        <a
+          href="https://discourse.aws.cloud.mov.ai/"
+          target="_blank"
+          rel="noreferrer"
+          className={classes.link}
+        >
+          <ChromeReaderModeIcon className={classes.linkIcon} />
+
+          {t("Forum")}
+        </a>
+        {/* <div className={classes.link} onClick={handleOpenAppConfig}>
           <BuildIcon className={classes.linkIcon} />
           {t("App Configuration")}
-        </div>
+        </div> */}
       </div>
     </Paper>
   );