+++ src/plugins/views/editors/Callback/Callback.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
+import { useTranslation } from "react-i18next";
 import { makeStyles, useTheme } from "@material-ui/core/styles";
 import { MonacoCodeEditor } from "@mov-ai/mov-fe-lib-code-editor";
 import { PLUGINS } from "./../../../../utils/Constants";
@@ -32,6 +33,7 @@ const Callback = (props, ref) => {
   // Style Hooks
   const classes = useStyles();
   const theme = useTheme();
+  const { t } = useTranslation();
 
   //========================================================================================
   /*                                                                                      *
@@ -41,15 +43,17 @@ const Callback = (props, ref) => {
 
   const renderRightMenu = React.useCallback(() => {
     const menuName = `${id}-detail-menu`;
+    const menuTitle = t("Callback Details Menu");
     // add bookmark
     call(PLUGINS.RIGHT_DRAWER.NAME, PLUGINS.RIGHT_DRAWER.CALL.SET_BOOKMARK, {
       [menuName]: {
         icon: <InfoIcon></InfoIcon>,
         name: menuName,
+        title: menuTitle,
         view: <Menu id={id} call={call} name={name} scope={scope} />
       }
     });
-  }, [call, id, name, scope]);
+  }, [call, id, name, scope, t]);
 
   usePluginMethods(ref, {
     renderRightMenu