+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # latest
 
+**Flow Editor**
+
+- [FP-1286 - Feature/fp 1286 port flow editor ](https://movai.atlassian.net/browse/FP-1286)
+
 **Home Tab**
 
+- [FP-1086 - Add branding in our ide](https://movai.atlassian.net/browse/FP-1086)
 - [FP-1346 - Integrate Home Tab in the IDE-CE](https://movai.atlassian.net/browse/FP-1346)
+- [FP-1457 - HomeTab adjustments](https://movai.atlassian.net/browse/FP-1457)